+++ tests/data/preview_context_managers/targeting_py39.py
@@ -84,7 +84,7 @@ with (
     # First comment.
     new_new_new1() as cm1,
     # Second comment.
-    new_new_new2()
+    new_new_new2(),
     # Last comment.
 ):
     pass